requirements.txt
certifi==2024.8.30
charset-normalizer==2.1.1
click==8.1.7
Flask==2.2.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
python-dotenv==0.21.0
requests==2.28.1
urllib3==1.26.20
Werkzeug==2.2.3
